+++ cipher/utils.h
@@ -46,7 +46,7 @@ inline Integer str_to_int(string str, int party) {
     return res;
 }
 
-inline void char_to_uint32(uint32_t* res, const char* in, uint64_t len) {
+inline void char_to_uint32(uint32_t* res, const char* in, size_t len) {
     for (int i = 0; i < len / 4; i++) {
     }
 }
@@ -105,8 +105,8 @@ inline void print_hex_32(Integer* s, int len) {
     cout << int_to_hex(outhex) << endl;
 }
 
-inline void intvec_to_int(Integer& out, Integer* in, uint64_t len) {
-    uint64_t s = in[0].size();
+inline void intvec_to_int(Integer& out, Integer* in, size_t len) {
+    size_t s = in[0].size();
     out = Integer(s * len, 0, PUBLIC);
     Integer tmp = Integer(s * len, 0, PUBLIC);
     for (int i = 0; i < len; i++) {
@@ -115,17 +115,17 @@ inline void intvec_to_int(Integer& out, Integer* in, uint64_t len) {
     }
 }
 
-inline void concat(Integer& res, const Integer* in, uint64_t len) {
+inline void concat(Integer& res, const Integer* in, size_t len) {
     for (int i = 0; i < len; i++)
         res.bits.insert(res.bits.begin(), in[i].bits.begin(), in[i].bits.end());
 }
 
-inline void reverse_concat(Integer& res, const Integer* in, uint64_t len) {
+inline void reverse_concat(Integer& res, const Integer* in, size_t len) {
     for (int i = 0; i < len; i++)
         res.bits.insert(res.bits.end(), in[i].bits.begin(), in[i].bits.end());
 }
 
-inline void move_concat(Integer& res, const Integer* in, uint64_t len) {
+inline void move_concat(Integer& res, const Integer* in, size_t len) {
     for (int i = 0; i < len; i++)
         res.bits.insert(res.bits.begin(), make_move_iterator(in[i].bits.begin()),
                         make_move_iterator(in[i].bits.end()));
@@ -210,7 +210,7 @@ inline block invBlock(block a) {
     return res;
 }
 
-inline block ghash(block h, block* x, uint64_t m) {
+inline block ghash(block h, block* x, size_t m) {
     block y = zero_block;
     for (int i = 0; i < m; i++) {
         y = mulBlock((y ^ x[i]), h);
@@ -257,7 +257,7 @@ inline block integer_to_block(Integer& in) {
 }
 
 // Transfer gc share into xor share.
-inline void integer_to_block(block* out, Integer* in, uint64_t len) {
+inline void integer_to_block(block* out, Integer* in, size_t len) {
     for (int i = 0; i < len; i++)
         out[i] = integer_to_block(in[i]);
 }
@@ -278,7 +278,7 @@ inline void integer_to_chars(unsigned char* out, Integer& in) {
     Integer ins(in);
     reverse(ins.bits.begin(), ins.bits.end());
     for (int i = 0; i < in.size(); i += 8) {
-        uint64_t tmp = 0;
+        size_t tmp = 0;
         for (int j = 0; j < 8; j++) {
             if (getLSB(ins.bits[i + j].bit)) {
                 tmp ^= (1 << (7 - j));
@@ -288,7 +288,7 @@ inline void integer_to_chars(unsigned char* out, Integer& in) {
     }
 }
 
-inline void block_to_hex(unsigned char* out, const block* in, uint64_t len) {
+inline void block_to_hex(unsigned char* out, const block* in, size_t len) {
     block* ins = new block[len];
     memcpy(ins, in, len * 16);
 
@@ -300,7 +300,7 @@ inline void block_to_hex(unsigned char* out, const block* in, uint64_t len) {
     delete[] ins;
 }
 
-inline void hex_to_block(block* out, const unsigned char* in, uint64_t len) {
+inline void hex_to_block(block* out, const unsigned char* in, size_t len) {
     if (len % 16 != 0)
         error("the length of the bytes is incorrect!\n");
     unsigned char* ins = new unsigned char[len];
